refactor(api): isolate project OctoBus mapping
